dummy utils整理
